Reduce chances of log files being overwritten
See https://bugs.mojang.com/browse/MC-100524 for more info.  This is a
bug that also affects vanilla logs, but since I can fix it for my plugin
I think I should.  The fix proposed in MC-100524 will probably also be
added to spigot, so it's not like it's just going to be here.

1000 means at max 1000 log files.  That should be enough for anyone (it
would require the server restarting once every 1.5 minutes for it to
reach that point).